condituation of integrating with Cognito, obtained Identity credentials, should be able to interact with a backend after this
